# DartsApi/DartsApi/Models/DTO/ThrowDto.cs
namespace DartsApi.Models.DTO
{
    public class ThrowDto
    {
        public int MatchId { get; set; }
        public int Throw1 { get; set; }
        public int Throw2 { get; set; }
        public int Throw3 { get; set; }
        public int SpelerId { get; set; }
        public DateTime Datum { get; set; }
    }
}