# DartsApi/DartsApi/Models/DTO/TournamentMatchDto.cs
namespace DartsApi.Models.DTO
{
    public class TournamentMatchDto
    {
        public int TournamentId { get; set; }
        public int MatchId { get; set; }
        public string Round {  get; set; }
    }
}